Don't add shadow-v0 to sanitizer script spec since it's not used in AMP
AMP shadow-v0 runtime is used in app shell context when serving the outer shell.
Since the outer shell is not an AMP document, it doesn't go through the sanitizer.

The inner shell, on the other hand, is always served with the regular v0 runtime
and so the script spec doesn't have to be modified in any way.

For these reasons, the code that adds shadow-v0 script to the allowed AMP tags list
is never executed and can be removed.